Class ZinbModel

Objects of this class store all the values needed to work with a zero-inflated negative binomial (ZINB) model, as described in the vignette. They contain all information to fit a model by penalized maximum likelihood or simulate data from a model.

# S4 method for ZinbModel
show(object)

# S4 method for ZinbModel
nSamples(x)

# S4 method for ZinbModel
nFeatures(x)

# S4 method for ZinbModel
nFactors(x)

# S4 method for ZinbModel
getX_mu(object, intercept = TRUE)

# S4 method for ZinbModel
getX_pi(object, intercept = TRUE)

# S4 method for ZinbModel
getV_mu(object, intercept = TRUE)

# S4 method for ZinbModel
getV_pi(object, intercept = TRUE)

# S4 method for ZinbModel
getLogMu(object)

# S4 method for ZinbModel
getMu(object)

# S4 method for ZinbModel
getLogitPi(object)

# S4 method for ZinbModel
getPi(object)

# S4 method for ZinbModel
getZeta(object)

# S4 method for ZinbModel
getPhi(object)

# S4 method for ZinbModel
getTheta(object)

# S4 method for ZinbModel
getEpsilon_beta_mu(object)

# S4 method for ZinbModel
getEpsilon_gamma_mu(object)

# S4 method for ZinbModel
getEpsilon_beta_pi(object)

# S4 method for ZinbModel
getEpsilon_gamma_pi(object)

# S4 method for ZinbModel
getEpsilon_W(object)

# S4 method for ZinbModel
getEpsilon_alpha(object)

# S4 method for ZinbModel
getEpsilon_zeta(object)

# S4 method for ZinbModel
getW(object)

# S4 method for ZinbModel
getBeta_mu(object)

# S4 method for ZinbModel
getBeta_pi(object)

# S4 method for ZinbModel
getGamma_mu(object)

# S4 method for ZinbModel
getGamma_pi(object)

# S4 method for ZinbModel
getAlpha_mu(object)

# S4 method for ZinbModel
getAlpha_pi(object)

Arguments

object
an object of class ZinbModel.

x
an object of class ZinbModel.

intercept
logical. Whether to return the intercept (ignored if the design matrix has no intercept). Default TRUE

Value

nSamples returns the number of samples; nFeatures returns the number of features; nFactors returns the number of latent factors.

Details

For the full description of the model see the model vignette. Internally, the slots are checked so that the matrices are of the appropriate dimensions: in particular, X, O_mu, O_pi, and W need to have n rows, V needs to have J rows, zeta must be of length J.

Methods (by generic)

show: show useful info on the object.

nSamples: returns the number of samples.

nFeatures: returns the number of features.

nFactors: returns the number of latent factors.

getX_mu: returns the sample-level design matrix for mu.

getX_pi: returns the sample-level design matrix for pi.

getV_mu: returns the gene-level design matrix for mu.

getV_pi: returns the sample-level design matrix for pi.

getLogMu: returns the logarithm of the mean of the non-zero component.

getMu: returns the mean of the non-zero component.

getLogitPi: returns the logit-probability of zero.

getPi: returns the probability of zero.

getZeta: returns the log of the inverse of the dispersion parameter.

getPhi: returns the dispersion parameter.

getTheta: returns the inverse of the dispersion parameter.

getEpsilon_beta_mu: returns the regularization parameters for beta_mu.

getEpsilon_gamma_mu: returns the regularization parameters for gamma_mu.

getEpsilon_beta_pi: returns the regularization parameters for beta_pi.

getEpsilon_gamma_pi: returns the regularization parameters for gamma_pi.

getEpsilon_W: returns the regularization parameters for W.

getEpsilon_alpha: returns the regularization parameters for alpha.

getEpsilon_zeta: returns the regularization parameters for zeta.

getW: returns the matrix W of inferred sample-level covariates.

getBeta_mu: returns the matrix beta_mu of inferred parameters.

getBeta_pi: returns the matrix beta_pi of inferred parameters.

getGamma_mu: returns the matrix gamma_mu of inferred parameters.

getGamma_pi: returns the matrix gamma_pi of inferred parameters.

getAlpha_mu: returns the matrix alpha_mu of inferred parameters.

getAlpha_pi: returns the matrix alpha_pi of inferred parameters.

Slots

X

matrix. The design matrix containing sample-level covariates, one sample per row.

V

matrix. The design matrix containing gene-level covariates, one gene per row.

O_mu

matrix. The offset matrix for mu.

O_pi

matrix. The offset matrix for pi.

which_X_mu

integer. Indeces of which columns of X to use in the regression of mu.

which_V_mu

integer. Indeces of which columns of V to use in the regression of mu.

which_X_pi

integer. Indeces of which columns of X to use in the regression of pi.

which_V_pi

integer. Indeces of which columns of V to use in the regression of pi.

X_mu_intercept

logical. TRUE if X_mu contains an intercept.

X_pi_intercept

logical. TRUE if X_pi contains an intercept.

V_mu_intercept

logical. TRUE if V_mu contains an intercept.

V_pi_intercept

logical. TRUE if V_pi contains an intercept.

W

matrix. The factors of sample-level latent factors.

beta_mu

matrix or NULL. The coefficients of X in the regression of mu.

gamma_mu

matrix or NULL. The coefficients of V in the regression of mu.

alpha_mu

matrix or NULL. The coefficients of W in the regression of mu.

beta_pi

matrix or NULL. The coefficients of X in the regression of pi.

gamma_pi

matrix or NULL. The coefficients of V in the regression of pi.

alpha_pi

matrix or NULL. The coefficients of W in the regression of pi.

zeta

numeric. A vector of log of inverse dispersion parameters.

epsilon_beta_mu

nonnegative scalar. Regularization parameter for beta_mu

epsilon_gamma_mu

nonnegative scalar. Regularization parameter for gamma_mu

epsilon_beta_pi

nonnegative scalar. Regularization parameter for beta_pi

epsilon_gamma_pi

nonnegative scalar. Regularization parameter for gamma_pi

epsilon_W

nonnegative scalar. Regularization parameter for W

epsilon_alpha

nonnegative scalar. Regularization parameter for alpha (both alpha_mu and alpha_pi)

epsilon_zeta

nonnegative scalar. Regularization parameter for zeta

epsilon_min_logit

scalar. Minimum regularization parameter for parameters of the logit model, including the intercept.