Parse ZINB regression model

Given the parameters of a ZINB regression model, this function parses the model and computes the vector of log(mu), logit(pi), and the dimensions of the different components of the vector of parameters. See zinb.loglik.regression for details of the ZINB regression model and its parameters.

zinb.regression.parseModel(alpha, A.mu, B.mu, C.mu, A.pi, B.pi, C.pi)

Arguments

alpha
the vectors of parameters c(a.mu, a.pi, b) concatenated

A.mu
matrix of the model (see above, default=empty)

B.mu
matrix of the model (see above, default=empty)

C.mu
matrix of the model (see above, default=zero)

A.pi
matrix of the model (see above, default=empty)

B.pi
matrix of the model (see above, default=empty)

C.pi
matrix of the model (see above, default=zero)

Value

A list with slots logMu, logitPi, dim.alpha (a vector of length 3 with the dimension of each of the vectors a.mu, a.pi and b in alpha), and start.alpha (a vector of length 3 with the starting indices of the 3 vectors in alpha)

See also